safari-600.7-branch/LayoutTests/css3/flexbox/overflow-keep-scrollpos-expected.txt

PASS document.querySelector('.overflow-auto').scrollTop is 50
